refactor + code splitting
