+++ src/mdxGlobalComponents.ts
@@ -17,7 +17,7 @@ import { ProductScreenshot } from './components/ProductScreenshot'
 import { ProductVideo } from './components/ProductVideo'
 import { StarRepoButton } from './components/StarRepoButton'
 import { CalloutBox } from './components/Docs/CalloutBox'
-import CodeReference from './components/CodeReference/CodeReference'
+import DemoReference from './components/DemoTour/DemoReference'
 
 export const shortcodes = {
     ArrayCTA,
@@ -39,5 +39,5 @@ export const shortcodes = {
     ProductVideo,
     StarRepoButton,
     CalloutBox,
-    CodeReference,
+    DemoReference,
 }
